Fix edit closes #75
Kind of weird fix, see comments in PhotoListFromIDs in backend.
What happened was that since we are now using ITKACL we dont use django users and therefore dont have users added to groups. I commented this out for now. I think this might be a permanent fix, since the user still have to be authenticated with itkacl and we should in theory be able to trust that.

Contact me for better explanation if needed. Point is. Should be fixed now, but still openening a new issue to check this when in beta just in case.